refactor: introduce command context management for Docker connections
