+++ core/cmd_connection/docker_attach.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/FMotalleb/crontab-go/abstraction"
 	"github.com/FMotalleb/crontab-go/config"
-	cmdutils "github.com/FMotalleb/crontab-go/core/cmd_connection/cmd_utils"
+	"github.com/FMotalleb/crontab-go/core/cmd_connection/command"
 	"github.com/FMotalleb/crontab-go/ctxutils"
 )
 
@@ -59,7 +59,7 @@ func NewDockerAttachConnection(log *logrus.Entry, conn *config.TaskConnection) (
 // Returns:
 // - An error if the preparation fails, otherwise nil.
 func (d *DockerAttachConnection) Prepare(ctx context.Context, task *config.Task) error {
-	cmdCtx := cmdutils.NewCtx(ctx, task.Env, d.log)
+	cmdCtx := command.NewCtx(ctx, task.Env, d.log)
 	d.ctx = ctx
 	// Specify the container ID or name
 	d.containerID = d.conn.ContainerName
